requirements.txt
langchain_openai>=0.0.5
python-dotenv>=1.0.0
browser_use>=0.1.0

langchain-openai~=0.3.1
pydantic~=2.11.4
selenium>=4.0.0
webdriver-manager>=3.8.0
undetected-chromedriver>=3.5.0